removes material-ui
