+++ client/app.js
@@ -10,7 +10,7 @@ import { me } from './store';
 /**
  * COMPONENT
  */
-class Routes extends Component {
+class App extends Component {
   componentDidMount() {
     this.props.loadInitialData();
   }
@@ -45,12 +45,12 @@ const mapDispatch = (dispatch) => {
   };
 };
 
-export default connect(mapState, mapDispatch)(Routes);
+export default connect(mapState, mapDispatch)(App);
 
 /**
  * PROP TYPES
  */
-Routes.propTypes = {
+App.propTypes = {
   loadInitialData: PropTypes.func.isRequired,
   isLoggedIn: PropTypes.bool.isRequired,
 };
