--- client/components/Colors.js
@@ -1,40 +1,0 @@
-export const primaryWhite = {
-  50: '#fefefe',
-  100: '#fefefe',
-  200: '#fdfdfd',
-  300: '#fcfcfc',
-  400: '#fbfbfb',
-  500: '#fafafa',
-  600: '#f9f9f9',
-  700: '#f9f9f9',
-  800: '#f8f8f8',
-  900: '#f6f6f6',
-  A100: '#ffffff',
-  A200: '#ffffff',
-  A400: '#ffffff',
-  A700: '#ffffff',
-  contrastDefaultColor: 'dark',
-};
-
-export const primaryWhite2 = {
-  50: '#ffffff',
-  100: '#ffffff',
-  200: '#ffffff',
-  300: '#ffffff',
-  400: '#ffffff',
-  500: '#ffffff',
-  600: '#ffffff',
-  700: '#ffffff',
-  800: '#ffffff',
-  900: '#ffffff',
-  A100: '#ffffff',
-  A200: '#ffffff',
-  A400: '#ffffff',
-  A700: '#ffffff',
-  contrastDefaultColor: 'dark',
-};
-
-exports.default = {
-  primaryWhite,
-  primaryWhite2,
-};
